requirements.txt
zulip>=0.9.0
trio>=0.22.0
PyYAML>=6.0